+++ fastapi_basekit/aio/sqlmodel/__init__.py
@@ -1,0 +1,9 @@
+from .controller.base import SQLModelBaseController
+from .repository.base import BaseRepository
+from .service.base import BaseService
+
+__all__ = [
+    "SQLModelBaseController",
+    "BaseRepository",
+    "BaseService",
+]