+++ fastapi_basekit/aio/sqlmodel/controller/__init__.py
@@ -1,0 +1,3 @@
+from .base import SQLModelBaseController
+
+__all__ = ["SQLModelBaseController"]